Add LIBINPUT_EVENT_POINTER_TOOL_UPDATE event type
This event would be emitted whenever a device with
LIBINPUT_DEVICE_CAP_STYLUS capabilities changes its tool type.

When the tablet reaches proximity, the tool/serial will be
updated to match the current physical stylus'. Likewise, when
the stylus moves out of proximity, the tool will be reset to
LIBINPUT_TOOL_NONE, so this event comes in pairs and can be used
to track stylus proximity.

Signed-off-by: Carlos Garnacho <carlosg@gnome.org>
Signed-off-by: Stephen Chandler Paul <thatslyude@gmail.com>