WIP: Add litest which reads/plays a device description and events from file
Running this test allows us to excersize libinput (and libevdev, as well) with
arbitrary device descriptions and event streams. The test is different from others
in that it doesn't make any assertions -- if the test completes without libinput
crashing, it succeeds. As such, this is intended to be used with something like
a fuzzer which can tweak good event streams into ones that go wildly or subtly
off the rails.